[diffusion] Move ComfyUI CUDA IPC helpers under runtime.distributed
Keep the handle/spill path with the other distributed transport code instead of at the runtime package root.
