+++ python/sglang/multimodal_gen/runtime/scheduler_client.py
@@ -26,7 +26,7 @@ from sglang.multimodal_gen.runtime.ipc_array import (
     is_local_endpoint,
     materialize_file_refs,
 )
-from sglang.multimodal_gen.runtime.ipc_cuda import (
+from sglang.multimodal_gen.runtime.distributed.ipc_cuda import (
     materialize_cuda_refs,
     release_retained_producer_tensors,
     spill_cuda_tensors,
